+++ test/Monotonic/Timer_test.rb
@@ -1,19 +1,19 @@
-require_relative '../../lib/Monotony/MonotonicTimer'
+require_relative '../../lib/Monotonic/Timer'
 
 require 'minitest/autorun'
 require 'minitest-spec-context'
 
-describe Monotony::MonotonicTimer do
+describe Monotonic::Timer do
   context "with a block" do
     it "the time spent in the block is returned as the value of the block" do
-      block_time = Monotony::MonotonicTimer.time do |timer|
+      block_time = Monotonic::Timer.time do |timer|
         sleep 3
       end
       expect(block_time.round).must_equal(3)
     end
 
     it "allows the timer to be started and stopped within the block" do
-      Monotony::MonotonicTimer.time do |timer|
+      Monotonic::Timer.time do |timer|
         sleep 1
         expect(timer.total_time.round).must_equal(1)
         timer.stop
@@ -30,7 +30,7 @@ describe Monotony::MonotonicTimer do
     end
 
     it "alters the block time if it is started and stopped within the block" do
-      block_time = Monotony::MonotonicTimer.time do |timer|
+      block_time = Monotonic::Timer.time do |timer|
         sleep 1
         timer.stop
         sleep 1
@@ -43,7 +43,7 @@ describe Monotony::MonotonicTimer do
 
   context "without a block" do
     it 'works' do
-      timer = Monotony::MonotonicTimer.new
+      timer = Monotonic::Timer.new
       timer.start
       sleep 1
       expect(timer.total_time.round).must_equal(1)
@@ -62,7 +62,7 @@ describe Monotony::MonotonicTimer do
 
   context "with a block on an timer instance" do
     it 'works' do
-      timer = Monotony::MonotonicTimer.new
+      timer = Monotonic::Timer.new
       time = timer.time do |timer|
         sleep 1
         expect(timer.total_time.round).must_equal(1)